Factor out some tool restore functionality into separate class
